requirements.txt
chronos-forecasting>=2.0.0
torch>=2.1.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
fastparquet>=2023.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
plotly>=5.18.0
yfinance>=0.2.28
prophet>=1.1.5
scikit-learn>=1.3.0
statsmodels>=0.14.0
scipy>=1.11.0
properscoring>=0.1
requests>=2.31.0
pyyaml>=6.0
